# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CV ATS Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WorkExperience {
  company: string;
  position: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  isCurrent: boolean;
  description: string;
}

export interface Education {
  institution: string;
  degree: string;
  field: string;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  isCurrent: boolean;
  /** @nullable */
  gpa?: string | null;
}

export interface ExtraSectionEntry {
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface ExtraSection {
  sectionTitle: string;
  entries: ExtraSectionEntry[];
}

export interface CVSummary {
  id: number;
  fullName: string;
  email: string;
  jobTitle: string;
  createdAt: string;
  updatedAt: string;
}

export interface Cv {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  jobTitle: string;
  summary: string;
  skills: string[];
  languages: string[];
  workExperience: WorkExperience[];
  education: Education[];
  extraSections: ExtraSection[];
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  cvLanguage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCVBody {
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  jobTitle: string;
  summary: string;
  skills: string[];
  languages: string[];
  workExperience: WorkExperience[];
  education: Education[];
  /** @nullable */
  extraSections?: ExtraSection[] | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  cvLanguage?: string | null;
}

export interface UpdateCVBody {
  fullName?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  jobTitle?: string;
  summary?: string;
  skills?: string[];
  languages?: string[];
  workExperience?: WorkExperience[];
  education?: Education[];
  /** @nullable */
  extraSections?: ExtraSection[] | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  cvLanguage?: string | null;
}

export interface CVHtml {
  html: string;
}