# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CV ATS Generator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all CVs
 */
export const ListCVsResponseItem = zod.object({
  id: zod.number(),
  fullName: zod.string(),
  email: zod.string(),
  jobTitle: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListCVsResponse = zod.array(ListCVsResponseItem);

/**
 * @summary Create a new CV
 */
export const CreateCVBody = zod.object({
  fullName: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  location: zod.string().nullish(),
  jobTitle: zod.string(),
  summary: zod.string(),
  skills: zod.array(zod.string()),
  languages: zod.array(zod.string()),
  workExperience: zod.array(
    zod.object({
      company: zod.string(),
      position: zod.string(),
      startDate: zod.string(),
      endDate: zod.string().nullish(),
      isCurrent: zod.boolean(),
      description: zod.string(),
    }),
  ),
  education: zod.array(
    zod.object({
      institution: zod.string(),
      degree: zod.string(),
      field: zod.string(),
      startDate: zod.string(),
      endDate: zod.string().nullish(),
      isCurrent: zod.boolean(),
      gpa: zod.string().nullish(),
    }),
  ),
  extraSections: zod
    .array(
      zod.object({
        sectionTitle: zod.string(),
        entries: zod.array(
          zod.object({
            title: zod.string(),
            subtitle: zod.string().nullish(),
            date: zod.string().nullish(),
            description: zod.string().nullish(),
          }),
        ),
      }),
    )
    .nullish(),
  linkedinUrl: zod.string().nullish(),
  portfolioUrl: zod.string().nullish(),
  cvLanguage: zod.string().nullish(),
});

/**
 * @summary Get a CV by ID
 */
export const GetCVParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCVResponse = zod.object({
  id: zod.number(),
  fullName: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  location: zod.string().nullish(),
  jobTitle: zod.string(),
  summary: zod.string(),
  skills: zod.array(zod.string()),
  languages: zod.array(zod.string()),
  workExperience: zod.array(
    zod.object({
      company: zod.string(),
      position: zod.string(),
      startDate: zod.string(),
      endDate: zod.string().nullish(),
      isCurrent: zod.boolean(),
      description: zod.string(),
    }),
  ),
  education: zod.array(
    zod.object({
      institution: zod.string(),
      degree: zod.string(),
      field: zod.string(),
      startDate: zod.string(),
      endDate: zod.string().nullish(),
      isCurrent: zod.boolean(),
      gpa: zod.string().nullish(),
    }),
  ),
  extraSections: zod.array(
    zod.object({
      sectionTitle: zod.string(),
      entries: zod.array(
        zod.object({
          title: zod.string(),
          subtitle: zod.string().nullish(),
          date: zod.string().nullish(),
          description: zod.string().nullish(),
        }),
      ),
    }),
  ),
  linkedinUrl: zod.string().nullish(),
  portfolioUrl: zod.string().nullish(),
  cvLanguage: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a CV
 */
export const UpdateCVParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCVBody = zod.object({
  fullName: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().nullish(),
  location: zod.string().nullish(),
  jobTitle: zod.string().optional(),
  summary: zod.string().optional(),
  skills: zod.array(zod.string()).optional(),
  languages: zod.array(zod.string()).optional(),
  workExperience: zod
    .array(
      zod.object({
        company: zod.string(),
        position: zod.string(),
        startDate: zod.string(),
        endDate: zod.string().nullish(),
        isCurrent: zod.boolean(),
        description: zod.string(),
      }),
    )
    .optional(),
  education: zod
    .array(
      zod.object({
        institution: zod.string(),
        degree: zod.string(),
        field: zod.string(),
        startDate: zod.string(),
        endDate: zod.string().nullish(),
        isCurrent: zod.boolean(),
        gpa: zod.string().nullish(),
      }),
    )
    .optional(),
  extraSections: zod
    .array(
      zod.object({
        sectionTitle: zod.string(),
        entries: zod.array(
          zod.object({
            title: zod.string(),
            subtitle: zod.string().nullish(),
            date: zod.string().nullish(),
            description: zod.string().nullish(),
          }),
        ),
      }),
    )
    .nullish(),
  linkedinUrl: zod.string().nullish(),
  portfolioUrl: zod.string().nullish(),
  cvLanguage: zod.string().nullish(),
});

export const UpdateCVResponse = zod.object({
  id: zod.number(),
  fullName: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  location: zod.string().nullish(),
  jobTitle: zod.string(),
  summary: zod.string(),
  skills: zod.array(zod.string()),
  languages: zod.array(zod.string()),
  workExperience: zod.array(
    zod.object({
      company: zod.string(),
      position: zod.string(),
      startDate: zod.string(),
      endDate: zod.string().nullish(),
      isCurrent: zod.boolean(),
      description: zod.string(),
    }),
  ),
  education: zod.array(
    zod.object({
      institution: zod.string(),
      degree: zod.string(),
      field: zod.string(),
      startDate: zod.string(),
      endDate: zod.string().nullish(),
      isCurrent: zod.boolean(),
      gpa: zod.string().nullish(),
    }),
  ),
  extraSections: zod.array(
    zod.object({
      sectionTitle: zod.string(),
      entries: zod.array(
        zod.object({
          title: zod.string(),
          subtitle: zod.string().nullish(),
          date: zod.string().nullish(),
          description: zod.string().nullish(),
        }),
      ),
    }),
  ),
  linkedinUrl: zod.string().nullish(),
  portfolioUrl: zod.string().nullish(),
  cvLanguage: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a CV
 */
export const DeleteCVParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get CV as HTML for printing/downloading
 */
export const GetCVHtmlParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCVHtmlResponse = zod.object({
  html: zod.string(),
});