# lib/api-zod/src/generated/types/createCVBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CV ATS Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { Education } from "./education";
import type { ExtraSection } from "./extraSection";
import type { WorkExperience } from "./workExperience";

export interface CreateCVBody {
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  jobTitle: string;
  summary: string;
  skills: string[];
  languages: string[];
  workExperience: WorkExperience[];
  education: Education[];
  /** @nullable */
  extraSections?: ExtraSection[] | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  cvLanguage?: string | null;
}